Update: refactor files into sub packages Add: LiteOperator With no ray dependency Add: async installation options
