+++ async_req.txt
@@ -1,0 +1,6 @@
+scikit-learn
+pydantic==1.9.1
+ray[default]==2.1.0
+redis
+protobuf~=3.19.0
+geomdl
