+++ rems/inputs/__init__.py
@@ -1,7 +1,5 @@
 from .InputBase import InputBase
 from .estop_callback import estop_callback
-from .KeyboardInput import KeyboardInput
-from .JoystickInput import JoystickInput
 from .FileCsvInput import FileCsvInput
 from .FileInput import FileInput
 from .HockInput import HockInput
