+++ rems/inputs/joy/__init__.py
@@ -1,0 +1,5 @@
+try:
+    from .JoystickInput import JoystickInput
+    from .JOYSTICK_KEYMAP import *
+except ImportError:
+    print("Keyboard Input requires 'pip install rems[keyboard]")
