+++ rems/inputs/key/__init__.py
@@ -1,0 +1,6 @@
+try:
+    from .KeyboardInput import KeyboardInput
+    from .KEYBOARD_KEYMAP import *
+except ImportError:
+    print("Keyboard Input requires 'pip install rems[keyboard]")
+
