--- rems/inputs/map/__init__.py
@@ -1,2 +1,0 @@
-from .JOYSTICK_KEYMAP import *
-#from .KEYBOARD_KEYMAP import *
