+++ rems/sim_handler/SimActor.py
@@ -1,14 +1,10 @@
-import ray, time
-import numpy as np
-from rems.sim_handler.ray.autocounter import auto_count
-from rems.robots import RobotBase
-from rems.utils.tictoc import tictoc
 import logging
-from ray.util import inspect_serializability
+import numpy as np
+import time
+
 ROUND = 2
 DISPLAY_MAX_ARRAY_LENGTH = 5
 
-@ray.remote
 class SimActor:
     def __init__(self, robot, outputs):
         self.robot = robot
@@ -19,7 +15,7 @@ class SimActor:
         self.t = 0.0
         logging.getLogger().setLevel(logging.INFO)
 
-    def step_forward(self, inpt, t_sys, DT):
+    def step_forward(self, inpt, t_sys, DT, *args, **kwargs):
         st = time.perf_counter()
         state = None
         observe = None
@@ -34,7 +30,7 @@ class SimActor:
         dt_actual = time.perf_counter() - st
         return observe, state, info, dt_actual, self.t
 
-    def step(self, inpt, t_sys, DT):
+    def step(self, inpt, t_sys, DT, *args, **kwargs):
         if np.round(self.t - t_sys, ROUND) >= DT:
             return
         #set input here
@@ -57,7 +53,7 @@ class SimActor:
             self.robot.run.name,
             np.round(dt_actual, 5), np.round(t, ROUND), disp[0], disp[1], disp[2], info))
 
-    def set_DT(self, DT):
+    def set_DT(self, DT, *args, **kwargs):
         if self.robot.run.DT is None:
             self.robot.run.DT = DT
 
@@ -65,19 +61,19 @@ class SimActor:
         return getattr(self.robot, name)(*args, **kwargs)
 
     # for communication
-    def _get_variable(self, name):
+    def _get_variable(self, name, *args, **kwargs):
         return getattr(self.robot, name)
 
 
-    def _set_variable(self, name, val):
+    def _set_variable(self, name, val, *args, **kwargs):
         #eval("self.robot."+name+"= val")
         setattr(self.robot, name, val)
         #getattr(self.robot, name)
 
-    def get_robot(self):
+    def get_robot(self, *args, **kwargs):
         return self.robot
 
-    def make_outputs(self):
+    def make_outputs(self, *args, **kwargs):
         # right now make outputs can be called only once
         for out in self.outputs:
             if out is None:
