+++ rems/sim_handler/job_background/__init__.py
@@ -1,6 +1,5 @@
 from rems.sim_handler.job_background.JobHandlerBase import JobHandlerBase
 from .JobHandler import JobHandler
-from .RayJobHandler import RayJobHandler
 from .JobHandlerMP import JobHandlerMP
 from .job_return_type import job_return_type
 from .job_type import job_type
