+++ rems/sim_handler/ray/SimActorRay.py
@@ -1,0 +1,6 @@
+from rems.sim_handler.SimActor import SimActor
+import ray
+
+@ray.remote
+class SimActorRay(SimActor):
+    pass
