+++ rems/sim_handler/ray/__init__.py
@@ -1,6 +1,6 @@
-from .ProcessActor import ProcessActor
-from .SimActor import SimActor
 from .autocounter import auto_count
 from .RobotRayWrapper import RobotRayWrapper
 from .RayWrapper import RayWrapper
 from .ActorWrapper import ActorWrapper
+from .RayJobHandler import RayJobHandler
+from .SimActorRay import SimActorRay
