+++ requirements.txt
@@ -1,17 +1,7 @@
-pygame
-matplotlib==3.5.0
-pandas
-pynput
+matplotlib
 numpy
-pytest
 scipy
-scikit-learn
-pydantic==1.9.1
-ray[default]==2.1.0
-redis
-protobuf~=3.19.0
 plotly
 pillow
-geomdl
-websocket-client
 defdict
+pandas
