+++ setup.py
@@ -2,6 +2,15 @@ from setuptools import setup, find_packages
 import codecs
 import os.path
 
+import sys
+
+def get_extras():
+    extras = []
+    for arg in sys.argv:
+        if '[' in arg and ']' in arg:
+            extras.extend(arg[arg.find('[') + 1:arg.find(']')].split(','))
+    return extras
+
 
 def read(rel_path):
     here = os.path.abspath(os.path.dirname(__file__))
@@ -33,15 +42,25 @@ version, cmdclass = get_version_and_cmdclass('rems')
 with open('requirements.txt') as f:
     requirements = f.read().splitlines()
 
+with open('async_req.txt') as f:
+    async_req = f.read().splitlines()
+
 setup(
     name='rems',
-    version='0.2.0',
+    version='0.3.0',
     cmdclass=cmdclass,
     description=get_metadata('description'),
     author=get_metadata('author'),
     license='LGPLv3',
     packages=find_packages(include=['rems', 'rems.*']),
     install_requires=requirements,
+    extras_require={
+        'core': [],
+        'async': async_req,
+        'joystick': ['pygame'],
+        'keyboard': ['pynput'],
+        'full': async_req + ['pygame', 'pynput']
+    },
     classifiers=[
         'Development Status :: 4 - Beta',
         'Framework :: Robot Framework',
